Resolve "upgrade ChemSpectra for prediction and mass"

Closes #720

See merge request ComPlat/chemotion_ELN!953

+++ app/models/concerns/attachment_jcamp_aasm.rb
@@ -78,18 +78,18 @@ module AttachmentJcampAasm
   def init_aasm
     return unless idle?
     _, extname = extension_parts
-    %w[dx jdx jcamp].include?(extname.downcase) ? set_queueing : set_non_jcamp
+    %w[dx jdx jcamp mzml raw].include?(extname.downcase) ? set_queueing : set_non_jcamp
   end
 
   def require_peaks_generation? # rubocop:disable all
     return unless belong_to_analysis?
     typname, extname = extension_parts
     return if peaked? || edited?
-    return unless %w[dx jdx jcamp].include?(extname.downcase)
+    return unless %w[dx jdx jcamp mzml raw].include?(extname.downcase)
     is_peak_edit = %w[peak edit].include?(typname)
     return generate_img_only(typname) if is_peak_edit
-    generate_spectrum(true, false, false, false) if queueing? && !new_upload
-    generate_spectrum(true, true, false, false) if regenerating? && !new_upload
+    generate_spectrum(true, false) if queueing? && !new_upload
+    generate_spectrum(true, true) if regenerating? && !new_upload
   end
 
   def belong_to_analysis?
@@ -105,6 +105,8 @@ module AttachmentJcampProcess
   extend ActiveSupport::Concern
 
   def generate_att(meta_tmp, addon, toEdit = false, ext = nil) # rubocop:disable AbcSize
+    return unless meta_tmp
+
     meta_filename = Chemotion::Jcamp::Gen.filename(filename_parts, addon, ext)
     content_type = ext == 'png' ? 'image/png' : 'application/octet-stream'
     att = Attachment.new(
@@ -128,36 +130,58 @@ module AttachmentJcampProcess
   end
 
   def generate_jcamp_att(jcamp_tmp, addon, toEdit = false)
-    use_default_ext = nil
-    generate_att(jcamp_tmp, addon, toEdit, use_default_ext)
+    generate_att(jcamp_tmp, addon, toEdit, 'jdx')
+  end
+
+  def build_params(params = {})
+    _, extname = extension_parts
+    params[:mass] = attachable.root_element.molecule.exact_molecular_weight || 0.0
+    params[:ext] = extname.downcase
+    params
+  end
+
+  def update_prediction(params, ori_pred, spc_type)
+    return auto_infer(spc_type) if spc_type == 'MS'
+    decision = params[:keep_pred] ?
+      ori_pred.decision :
+      JSON.parse(params['predict'])
+    predictions.create(decision: decision)
   end
 
   def create_process(is_regen)
-    tmp_jcamp, tmp_img = Chemotion::Jcamp::Create.spectrum(abs_path, is_regen)
-    generate_jcamp_att(tmp_jcamp, 'peak')
+    params = build_params
+    tmp_jcamp, tmp_img, spc_type = Chemotion::Jcamp::Create.spectrum(
+      abs_path, is_regen, params
+    )
+    jcamp_att = generate_jcamp_att(tmp_jcamp, 'peak')
+    jcamp_att.auto_infer(spc_type)
     img_att = generate_img_att(tmp_img, 'peak')
     set_done
     delete_tmps([tmp_jcamp, tmp_img])
     delete_related_imgs(img_att)
     delete_edit_peak_after_done
+    jcamp_att
   end
 
-  def edit_process(is_regen, peaks, shift)
-    tmp_jcamp, tmp_img = Chemotion::Jcamp::Create.spectrum(
-      abs_path, is_regen, peaks, shift
+  def edit_process(is_regen, orig_params)
+    params = build_params(orig_params)
+    tmp_jcamp, tmp_img, spc_type = Chemotion::Jcamp::Create.spectrum(
+      abs_path, is_regen, params
     )
-    generate_jcamp_att(tmp_jcamp, 'edit', true)
+    jcamp_att = generate_jcamp_att(tmp_jcamp, 'edit', true)
+    jcamp_att.update_prediction(params, predictions[0], spc_type)
     img_att = generate_img_att(tmp_img, 'edit', true)
     set_backup
     delete_tmps([tmp_jcamp, tmp_img])
     delete_related_imgs(img_att)
     delete_edit_peak_after_done
+    jcamp_att
   end
 
   def generate_spectrum(
-    is_create = false, is_regen = false, peaks = false, shift = false
+    is_create = false, is_regen = false, params = {}
   )
-    is_create ? create_process(is_regen) : edit_process(is_regen, peaks, shift)
+    is_create ? create_process(is_regen) : edit_process(is_regen, params)
   rescue
     set_failure
     Rails.logger.info('**** Jcamp Peaks Generation fails ***')
@@ -165,6 +189,7 @@ module AttachmentJcampProcess
 
   def delete_tmps(tmp_arr)
     tmp_arr.each do |tmp|
+      next unless tmp
       tmp.close
       tmp.unlink
     end
@@ -172,7 +197,7 @@ module AttachmentJcampProcess
 
   def delete_edit_peak_after_done
     typname = extension_parts[0]
-    delete if %w[edit peak].include?(typname)
+    destroy if %w[edit peak].include?(typname)
   end
 
   def fname_wo_ext(target)
@@ -182,6 +207,8 @@ module AttachmentJcampProcess
   end
 
   def delete_related_imgs(img_att)
+    return unless img_att
+
     atts = Attachment.where(attachable_id: attachable_id)
     valid_name = fname_wo_ext(self)
     atts.each do |att|
@@ -203,5 +230,58 @@ module AttachmentJcampProcess
   rescue
     set_failure
     Rails.logger.info('**** Jcamp Image Generation fails ***')
+  end
+
+  def infer_base_on_type(t_molfile, params)
+    case params[:layout]
+    when 'IR'
+      spectrum = read_file
+      Tempfile.create('spectrum') do |t_spectrum|
+        t_spectrum.write(spectrum)
+        t_spectrum.rewind
+        Chemotion::Jcamp::Predict::Ir.exec(
+          t_molfile, t_spectrum
+        )
+      end
+    when 'MS'
+      spectrum = read_file
+      Tempfile.create('spectrum') do |t_spectrum|
+        t_spectrum.write(spectrum)
+        t_spectrum.rewind
+        Chemotion::Jcamp::Predict::MS.exec(
+          t_molfile, t_spectrum
+        )
+      end
+    else
+      Chemotion::Jcamp::Predict::NmrPeaksForm.exec(
+        t_molfile, params[:layout], params[:peaks], params[:shift]
+      )
+    end
+  end
+
+  def infer_with_molfile(params)
+    molfile = attachable.root_element.molecule.molfile
+    Tempfile.create('molfile') do |t_molfile|
+      t_molfile.write(molfile)
+      t_molfile.rewind
+      infer_base_on_type(t_molfile, params)
+    end
+  end
+
+  def infer_spectrum(params)
+    target = infer_with_molfile(params)
+    return until target
+    predictions.destroy_all
+    predictions.create(decision: target)
+    target
+  end
+
+  def auto_infer(spc_type)
+    case spc_type
+    when 'INFRARED'
+      infer_spectrum({ layout: 'IR' })
+    when 'MS'
+      infer_spectrum({ layout: 'MS' })
+    end
   end
 end
